Initial start of notmuch web interface using web.py
This is intended to serve two purposes: first, to be a useful
web mail system for at least some users.  As long as it stays
in notmuch's contrib directory, it'll probably require being run
as the user whose mail will be shown, and auth will be Some Other
Server's Problem.  Anyone who wants to make it multi-user is
welcome to spin it off.

Second, it's meant to be a demo project using the Python bindings.
The old notmuch.py served that role; now this can.

The initial draft only supports enough to prove it can work:
a main page with tags, and a search page that shows results by
subject and author.  To be worth using, it ought to offer

  - safe escaping of values from messages
  - better formatting
  - ability to show each individual message
  - reply functionality, probably with mailto: links